+++ packages/plugin-gallery/web/src/components/gallery-controls/gallery-items-sortable.js
@@ -520,7 +520,8 @@ export class SortableComponent extends Component {
 
   onUpdateImage = metadata => {
     const { editedImage } = this.state;
-    this.setState({ editedImage, metadata: { ...editedImage.metadata, ...metadata } });
+    editedImage.metadata = { ...editedImage.metadata, ...metadata };
+    this.setState({ editedImage });
   };
 
   onDeleteImage = () => {