+++ CHANGELOG.md
@@ -13,6 +13,10 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `gallery`
+  - [#672](https://github.com/wix-incubator/rich-content/pull/672) gallery images settings frozen
+  
 </details>
 <hr/>